requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
distlib==0.3.1
docutils==0.16
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gTTS==2.2.2
identify==1.6.1
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.5
nodeenv==1.5.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pefile==2019.4.18
Pillow==8.1.1
pluggy==0.13.1
pre-commit==2.10.1
py==1.10.0
pycodestyle==2.6.0
Pygments==2.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyjokes==0.6.0
pylint==2.6.2
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tox==3.22.0
tqdm==4.58.0
translate==3.5.0
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
